app/requirements.txt
luma.oled
timeloop
requests
xmltodict
RPi.GPIO
spidev
flask